refactor: rename task response metadata sol
