+++ crates/task-processor/src/lib.rs
@@ -28,8 +28,6 @@ use tracing::info;
 
 /// Task processor error
 pub mod error;
-/// Task response metadata
-pub mod response_metadata;
 /// Task
 pub mod task;
 /// Task manager trait
@@ -38,6 +36,8 @@ pub mod task_manager;
 pub mod task_processor;
 /// Task response
 pub mod task_response;
+/// Task response metadata
+pub mod task_response_metadata_sol;
 
 type TaskResponsesMap<O> = HashMap<u32, HashMap<TaskResponseDigest, TaskResponse<O>>>;
